pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "superego-mcp"
version = "0.1.0"
description = "Intelligent tool request interception for AI agents"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Superego MCP Team" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.0.0",           # MCP server framework with sampling support
    "pydantic>=2.0.0",          # Data validation and domain models
    "pyyaml>=6.0",              # Configuration file parsing
    "watchfiles>=0.20.0",       # File system monitoring for hot-reload
    "jinja2>=3.1.0",            # Secure prompt templating
    "httpx>=0.25.0",            # HTTP client for AI services
    "structlog>=23.0.0",        # Structured logging
    "psutil>=5.9.0",            # System metrics for health checks
]

[project.optional-dependencies]
demo = [
    "fast-agent-mcp>=0.1.0",    # Demo client for sampling testing
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
superego-mcp = "superego_mcp.main:cli_main"

[project.urls]
Homepage = "https://github.com/your-org/superego-mcp"
Repository = "https://github.com/your-org/superego-mcp"
Issues = "https://github.com/your-org/superego-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/superego_mcp"]

[tool.hatch.version]
path = "src/superego_mcp/__init__.py"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "fastmcp.*",
    "fast_agent_mcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=superego_mcp",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]